ADD: all neccesary files, structure is done, here is the first version of the program

+++ sem01/lab_05/headers/queue.h
@@ -5,7 +5,7 @@
 
 typedef struct node_t
 {
-    double data;
+    int data;
     struct node_t *next;
 } node_t;
